[2/2] Rename DownloadController to RepositoryController
Rename module
